requirements.txt
streamlit
mysql.connector
openai.OpenAI
st_chat_message
mysql.connector
